requirements.txt
questionary==1.5.2
tqdm==4.64.0
gspread==3.6.0
simple_youtube_api==0.2.8
google_api_core==1.31.5
youtube_dl==2020.9.20
gdown==4.4.0
pytube3==9.6.4
requests==2.24.0
PyDrive==1.3.1
isodate==0.6.1
Google==3.0.0
google_api_python_client==2.47.0
pytube==12.0.0
PyYAML==6.0
unicode_slugify==0.1.5
